separate out invalid.py
